+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.1.1 - 2016-10-26
+### Added
+- Attribute to set concurrency for sentry worker (default is number of processors in machine).
+- Attribute to set 'numprocs' for worker supervisor config (default is 1).
+- Default Redis optimization options according to https://docs.sentry.io/server/performance/.
+
 ## 0.1.0 - 2016-10-26
 ### Added
 - Initial working version of the cookbook